chore: Move slice files to the store directory.
